# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/collections": {
            "get": {
                "description": "List all collection configurations",
                "produces": [
                    "application/json"
                ],
                "summary": "List all collections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/collections.Collection"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new collection configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create collection",
                "parameters": [
                    {
                        "description": "Collection data",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collections.Collection"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/collections.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/collections/{name}": {
            "get": {
                "description": "Get a collection configuration by name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collections.Collection"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a collection configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Collection data",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collections.Collection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collections.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a collection configuration",
                "summary": "Delete collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/collections/{name}/destinations": {
            "get": {
                "description": "Get destinations for a collection",
                "produces": [
                    "application/json"
                ],
                "summary": "Get collection destinations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/collections.DestinationConfig"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Replace destinations for a collection (stream_enabled must be true)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update collection destinations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Destinations data",
                        "name": "destinations",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/collections.DestinationConfig"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/collections.DestinationConfig"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "collections.Collection": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "collection_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deletion_protection": {
                    "type": "boolean"
                },
                "destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/collections.DestinationConfig"
                    }
                },
                "old_image": {
                    "type": "boolean"
                },
                "primary_key": {
                    "type": "string"
                },
                "sort_key": {
                    "type": "string"
                },
                "stream_enabled": {
                    "type": "boolean"
                },
                "ttl_attribute": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "collections.DestinationConfig": {
            "type": "object",
            "properties": {
                "bearer_token": {
                    "description": "HTTP auth token or Meilisearch API key",
                    "type": "string"
                },
                "endpoint": {
                    "description": "HTTP URL or Meilisearch host or EventBridge event-bus name",
                    "type": "string"
                },
                "event_bus_name": {
                    "description": "EventBridge event bus name",
                    "type": "string"
                },
                "event_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter_criteria": {
                    "$ref": "#/definitions/collections.FilterCriteria"
                },
                "index_name": {
                    "description": "Meilisearch-specific",
                    "type": "string"
                },
                "region": {
                    "description": "EventBridge-specific",
                    "type": "string"
                },
                "source": {
                    "description": "Event source (default: \"conduit-mongodb\")",
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "collections.FilterCondition": {
            "type": "object",
            "properties": {
                "anything-but": {},
                "exists": {
                    "type": "boolean"
                },
                "numeric": {
                    "type": "array",
                    "items": {}
                },
                "prefix": {
                    "type": "string"
                },
                "suffix": {
                    "type": "string"
                }
            }
        },
        "collections.FilterCriteria": {
            "type": "object",
            "properties": {
                "new_image": {
                    "$ref": "#/definitions/collections.ImageFilter"
                },
                "old_image": {
                    "$ref": "#/definitions/collections.ImageFilter"
                }
            }
        },
        "collections.ImageFilter": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/collections.FilterCondition"
            }
        }
    }
}